[refactor] 채팅방 참여자 수를 participant 집계 기반으로 전환

+++ src/main/java/_ganzi/codoc/chat/dto/ChatRoomListQueryResult.java
@@ -3,4 +3,4 @@ package _ganzi.codoc.chat.dto;
 import java.time.Instant;
 
 public record ChatRoomListQueryResult(
-        Long roomId, String title, boolean hasPassword, int participantCount, Instant lastMessageAt) {}
+        Long roomId, String title, boolean hasPassword, long participantCount, Instant lastMessageAt) {}
